refactor: extract types into their own modules
